Move `Rate_Limit_Exception`
